feat(vote): implement user voting
- validate input
- ensure only authenticated user can vote
- ensure user can only vote once
- ensure that the candidate and office exists
