+++ server/middlewares/AdminValidator.js
@@ -3,6 +3,7 @@ import pool from '../config/connection';
 
 
 const { extractErrors } = Helpers;
+
 /**
  * Defines methods for validating Admin functions
  *
@@ -15,6 +16,26 @@ class AdminValidator {
       .isNumeric()
       .withMessage('The office must be a number');
 
+    const errors = req.validationErrors();
+    if (errors) {
+      return res.status(400).json({
+        status: 400,
+        errors: extractErrors(errors),
+      });
+    }
+    next();
+  }
+
+  static validateVote(req, res, next) {
+    req.check('office', 'The aspirant\'s office is required')
+      .notEmpty().trim()
+      .isNumeric()
+      .withMessage('The office must be a number');
+
+    req.check('candidate', 'Select the candiadte to vote for')
+      .notEmpty().trim()
+      .isNumeric()
+      .withMessage('The candidate ID must be a number');
     const errors = req.validationErrors();
     if (errors) {
       return res.status(400).json({
@@ -83,7 +104,7 @@ class AdminValidator {
     } finally {
       await client.release();
     }
-    return next();
+    next();
   }
 
   /**
@@ -111,7 +132,7 @@ class AdminValidator {
         error: 'The user ID is required',
       });
     }
-    return next();
+    next();
   }
 
   /**
@@ -143,6 +164,50 @@ class AdminValidator {
       await client.release();
     }
     req.body.userId = userId;
+    next();
+  }
+
+  static async checkIfUserHasVoted(req, res, next) {
+    /** No validation is done here because the request is passed down by Authenticator */
+    const officeId = req.body.office;
+    let voterId;
+    const client = await pool.connect();
+    try {
+      const decodedToken = req.body.token;
+      voterId = decodedToken.id;
+      if (!voterId) {
+        return res.status(401).json({ status: 401, error: 'Unauthorized voter, You must login to vote' });
+      }
+      const sqlQuery = { text: 'SELECT * FROM votes WHERE voter = $1 AND office = $2',
+        values: [voterId, officeId] };
+      const user = await client.query(sqlQuery);
+      if (user.rowCount) {
+        return res.status(409).json({ status: 409, error: 'you have already voted for this office' });
+      }
+    } catch (err) {
+      return;
+    } finally { await client.release(); }
+    req.body.voter = voterId;
+    next();
+  }
+
+  static async checkIfCandidateExists(req, res, next) {
+    const { candidate } = req.body;
+    const client = await pool.connect();
+    try {
+      const sqlQuery = { text: 'SELECT * FROM candidates WHERE id = $1', values: [candidate] };
+      const candidateExists = await client.query(sqlQuery);
+      if (candidateExists.rowCount === 0) {
+        return res.status(404).json({
+          status: 404,
+          error: 'The candidate does not exist',
+        });
+      }
+    } catch (err) {
+      return;
+    } finally {
+      await client.release();
+    }
     next();
   }
 }
